Finish working version of python tutorial.
Split Simulation environments into base and custom files.
Fix bugs in types for reward functions. Add typing to
gym algorithms file.
