+++ acnportal/acnsim/gym_acnsim/__init__.py
@@ -1,4 +1,3 @@
-from . import reward_functions
 from importlib.util import find_spec
 if find_spec("gym") is not None:
     from gym.envs.registration import register
@@ -23,5 +22,6 @@ if find_spec("gym") is not None:
         entry_point='acnportal.acnsim.gym_acnsim.envs:'
                     'make_rebuilding_default_sim_env',
     )
-    from . import observation
+    from .envs import action_spaces, observation, reward_functions
+    from . import reward_functions
 del find_spec
