+++ acnportal/acnsim/gym_acnsim/envs/__init__.py
@@ -1,8 +1,11 @@
 from importlib.util import find_spec
 if find_spec("gym") is not None:
-    from .sim_prototype_env import BaseSimEnv
-    from .sim_prototype_env import CustomSimEnv
-    from .sim_prototype_env import RebuildingEnv
-    from .sim_prototype_env import make_default_sim_env
-    from .sim_prototype_env import make_rebuilding_default_sim_env
+    from .base_env import BaseSimEnv
+    from .custom_envs import CustomSimEnv
+    from .custom_envs import RebuildingEnv
+    from .custom_envs import make_default_sim_env
+    from .custom_envs import make_rebuilding_default_sim_env
+    from .custom_envs import default_observation_objects
+    from .custom_envs import default_action_object
+    from .custom_envs import default_reward_functions
 del find_spec
