+++ acnportal/algorithms/__init__.py
@@ -3,6 +3,7 @@ from .uncontrolled_charging import UncontrolledCharging
 from .sorted_algorithms import *
 from importlib.util import find_spec
 if find_spec("gym") is not None:
+    from .open_ai_algorithms import GymBaseAlgorithm
     from .open_ai_algorithms import GymTrainedAlgorithm
     from .open_ai_algorithms import GymTrainingAlgorithm
     from .open_ai_algorithms import SimRLModelWrapper
