+++ acnportal/algorithms/base_algorithm.py
@@ -25,7 +25,7 @@ class BaseAlgorithm:
         if self._interface is not None:
             return self._interface
         else:
-            raise ValueError('No interface has been registered yet. Please call register_interface prior to using the'
+            raise ValueError('No interface has been registered yet. Please call register_interface prior to using the '
                              'algorithm.')
 
     def register_interface(self, interface):
